server/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==2.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
watchdog==6.0.0
Werkzeug==3.1.3
